LayoutTests/http/tests/security/xss-DENIED-contentWindow-eval-expected.txt

CONSOLE MESSAGE: line 1: TypeError: Type error
This test passes if alert() is not called. 
